- New typedef color_t (int representation of color, use this for 24-bit color, and byte for quake pallete). - Changed the name of RGBA_2_Int/Int_2_RGBA to RGBA_TO_COLOR/COLOR_TO_RGBA. - Changed all Draw-functions that used floats to use color_t instead. - Removed a few drawing functions that were not used anywhere.

+++ hud.h
@@ -59,8 +59,8 @@ typedef struct hud_s
     cvar_t *show;						// Show cvar.
     cvar_t *frame;						// Frame cvar.
 	cvar_t *frame_color;				// Frame color cvar.
-	float frame_color_cache[4];			// Cache for parsed frame color.
-	
+	byte frame_color_cache[4];			// Cache for parsed frame color.
+
 #if defined(FRAMEBUFFERS) && defined(GLQUAKE)
 	cvar_t *opacity;					// The overall opacity of the entire HUD element.
 #endif
